Ethereum out-channel (#208)
* first implm

* add second channel and rename

* add events to abstract class

* fix readme

* fix config script

* misc pr fixes and test improvements

* add basic channel test

* fix tests and erc20

* remove relayer changes

+++ ethereum/contracts/Scale.sol
@@ -2,16 +2,11 @@
 pragma solidity >=0.6.2;
 
 contract Scale {
-
     // Decodes a SCALE encoded uint256 by converting bytes (bid endian) to little endian format
-    function decodeUint256(bytes memory data)
-        public
-        pure
-        returns (uint256)
-    {
+    function decodeUint256(bytes memory data) public pure returns (uint256) {
         uint256 number;
-        for(uint i = data.length; i > 0; i--) {
-            number = number + uint(uint8(data[i-1])) * (2 ** (8 * (i-1)));
+        for (uint256 i = data.length; i > 0; i--) {
+            number = number + uint256(uint8(data[i - 1])) * (2**(8 * (i - 1)));
         }
         return number;
     }
@@ -22,31 +17,38 @@ contract Scale {
         pure
         returns (uint256)
     {
-        uint8 b = readByteAtIndex(data, 0);           // read the first byte
-        uint8 mode = b & 3;                           // bitwise operation
+        uint8 b = readByteAtIndex(data, 0); // read the first byte
+        uint8 mode = b & 3; // bitwise operation
 
-        if (mode == 0) {                               // [0, 63]
-            return b >> 2;                            // right shift to remove mode bits
-        } else if (mode == 1) {                        // [64, 16383]
-            uint8 bb = readByteAtIndex(data, 1);      // read the second byte
-            uint64 r = bb;                            // convert to uint64
-            r <<= 6;                                  // multiply by * 2^6
-            r += b >> 2;                              // right shift to remove mode bits
+        if (mode == 0) {
+            // [0, 63]
+            return b >> 2; // right shift to remove mode bits
+        } else if (mode == 1) {
+            // [64, 16383]
+            uint8 bb = readByteAtIndex(data, 1); // read the second byte
+            uint64 r = bb; // convert to uint64
+            r <<= 6; // multiply by * 2^6
+            r += b >> 2; // right shift to remove mode bits
             return r;
-        } else if (mode == 2) {                        // [16384, 1073741823]
-            uint8 b2 = readByteAtIndex(data, 1);      // read the next 3 bytes
+        } else if (mode == 2) {
+            // [16384, 1073741823]
+            uint8 b2 = readByteAtIndex(data, 1); // read the next 3 bytes
             uint8 b3 = readByteAtIndex(data, 2);
             uint8 b4 = readByteAtIndex(data, 3);
 
-            uint32 x1 = uint32(b) | uint32(b2) << 8;  // convert to little endian
-            uint32 x2 = x1 | uint32(b3) << 16;
-            uint32 x3 = x2 | uint32(b4) << 24;
+            uint32 x1 = uint32(b) | (uint32(b2) << 8); // convert to little endian
+            uint32 x2 = x1 | (uint32(b3) << 16);
+            uint32 x3 = x2 | (uint32(b4) << 24);
 
-            x3 >>= 2;                                 // remove the last 2 mode bits
+            x3 >>= 2; // remove the last 2 mode bits
             return uint256(x3);
-        } else if (mode == 3) {                        // [1073741824, 4503599627370496]
-            uint8 l = b >> 2;                         // remove mode bits
-            require(l > 32, "Not supported: number cannot be greater than 32 bytes");
+        } else if (mode == 3) {
+            // [1073741824, 4503599627370496]
+            uint8 l = b >> 2; // remove mode bits
+            require(
+                l > 32,
+                "Not supported: number cannot be greater than 32 bytes"
+            );
         } else {
             revert("Code should be unreachable");
         }
